Rename `ticascii` to `ticli`.
This is required because `ticascii` can't import from `ticascii.players`, because that's a reference to itself.  It has to be able to do an absolute import like that because when `ticli.py` is made into a script, it will be moved into a local distribution "bin/" folder, so it can't rely on `ticascii/` being in the local directory but instead relies on the module `ticascii` being importable from within the environment.

The name `ticli` is carefully chosen as a play on the words "tic" and "CLI".